Add files via upload
These are notes on: Variables, strings, integers, floats, Booleans, concatenation, f-strings, math operators, subscripting, print(), input(), type() and various small program challenges. These small programs are the first step to understanding the basics of the above material. The final program is a tip calculator that encompasses most of this.

This program will be updated at a further date.